Extract config to a package
